# Conflicts:
#	src/main/kotlin/com/odtheking/odin/OdinMod.kt
#	src/main/kotlin/com/odtheking/odin/events/EventDispatcher.kt
#	src/main/kotlin/com/odtheking/odin/features/impl/dungeon/puzzlesolvers/TPMazeSolver.kt

+++ src/main/kotlin/com/odtheking/odin/clickgui/HudManager.kt
@@ -31,7 +31,7 @@ object HudManager : Screen(Component.literal("HUD Manager")) {
         }
 
         context.pose()?.pushMatrix()
-        val sf = mc.window.guiScale.toFloat()
+        val sf = mc.window.guiScale
         context.pose().scale(1f / sf, 1f / sf)
 
         for (hud in hudSettingsCache) {
